+++ frontend/src/components/ModelSelectionMenu.jsx
@@ -7,11 +7,13 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import ModelBubble from "./ModelBubble";
 import { useEffect, useState } from "react";
 import useFetchData from "../hooks/useFetchData";
+import useLoadModel from "../hooks/useLoadModel";
 
-const ModelSelectionMenu = ({ isOpen, setIsOpen, setModel, serverUrl }) => {
+const ModelSelectionMenu = ({ isOpen, setIsOpen, setModel, serverUrl, setModelLoading }) => {
   const url = `${serverUrl}/v1/models`;
   const [modelList, setModelList] = useState([]);
   const { data, loading, error } = useFetchData(url);
+  const { response, loading: modelLoading, loadModel } = useLoadModel(serverUrl);
 
   const handleOutsideClick = (e) => {
     if (e.target === e.currentTarget) setIsOpen(false);
@@ -23,6 +25,7 @@ const ModelSelectionMenu = ({ isOpen, setIsOpen, setModel, serverUrl }) => {
 
   const handleModelSelect = (model) => {
     setModel(model);
+    loadModel(model);
     setIsOpen(false);
   };
 
@@ -35,6 +38,11 @@ const ModelSelectionMenu = ({ isOpen, setIsOpen, setModel, serverUrl }) => {
     }
   }, [data, loading]);
 
+  useEffect(() => {
+    setModelLoading(modelLoading);
+  }
+    , [modelLoading]);
+  
   return (
     <div
       className={`